+++ Controllers/TareaController.cs
@@ -29,16 +29,17 @@ public class TareaController : Controller
     public IActionResult Create(int id)
     {
         if(!EstaLogeado()) return RedirectToRoute(new {controller = "Login", action="Index"});
-        var tarea = new CrearTareaViewModel();
+        var tarea = new CrearTareaViewModel(_usuarioRepository.GetAll());
         tarea.Id_tablero = id;
         return View(tarea); 
     }
 
     [HttpPost]
-    public IActionResult Create(Tarea tarea)
+    public IActionResult Create(CrearTareaViewModel t)
     {
         if(!EstaLogeado()) return RedirectToRoute(new {controller = "Login", action="Index"});
         if(!ModelState.IsValid) return RedirectToAction("Create");
+        Tarea tarea = new Tarea(t);
         _tareaRepository.Create(tarea.Id_tablero, tarea);
         return RedirectToAction("Index", new{id = tarea.Id_tablero});
     }
@@ -47,23 +48,43 @@ public class TareaController : Controller
     public IActionResult Update(int id)
     {  
         if(!EstaLogeado()) return RedirectToRoute(new {controller = "Login", action="Index"});
-        if(!ModelState.IsValid) return RedirectToAction("Update");
-        return View(new ActualizarTareaViewModel(_tareaRepository.GetById(id), _usuarioRepository.GetAll() ));
+        Tarea nuevaTarea = _tareaRepository.GetById(id);
+        if (IsAdmin() || _tableroRepository.GetById(nuevaTarea.Id_tablero).IdUsuarioPropietario == Convert.ToInt32(HttpContext.Session.GetString("Id"))) // solo admin y propietario de la tarea
+        {
+            return View(new ActualizarTareaViewModel(nuevaTarea, _usuarioRepository.GetAll()));
+        }else
+        {
+            return RedirectToAction("Error");
+        }
     }
 
     [HttpPost]
-    public IActionResult Update(ActualizarTareaViewModel tarea) {
+    public IActionResult Update(ActualizarTareaViewModel t) {
+        if(!ModelState.IsValid) return RedirectToAction("Update");
         if(!EstaLogeado()) return RedirectToRoute(new {controller = "Login", action="Index"});
-        Tarea tarea1 = new Tarea(tarea);
-        _tareaRepository.Update(tarea1.Id, tarea1);
-        return RedirectToAction("Index", new{id = tarea1.Id_tablero});
+        if(IsAdmin() || _tableroRepository.GetById(t.Id_tablero).IdUsuarioPropietario == Convert.ToInt32(HttpContext.Session.GetString("Id"))){
+            Tarea tarea = new Tarea(t);
+            _tareaRepository.Update(tarea.Id, tarea);
+            return RedirectToAction("Index", new{id = tarea.Id_tablero});
+        }else
+        {
+             return RedirectToAction("Error");
+        }
     }
 
     [HttpGet]
     public IActionResult Delete(int id) {
         if(!EstaLogeado()) return RedirectToRoute(new {controller = "Login", action="Index"});
-        _tareaRepository.Remove(id);
-        return RedirectToAction("Index");
+
+        Tarea tareaEliminar = _tareaRepository.GetById(id);
+        if (IsAdmin() || _tableroRepository.GetById(tareaEliminar.Id_tablero).IdUsuarioPropietario == Convert.ToInt32(HttpContext.Session.GetString("Id")))
+        {
+            _tareaRepository.Remove(id);
+            return RedirectToAction("Index");   
+        }else
+        {
+            return RedirectToAction("Error");
+        }
     }
 
     private bool IsAdmin() => HttpContext.Session.GetString("NivelDeAcceso") == enumRol.admin.ToString();
